// src/api/model/create-workspace-request-dto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */

export interface CreateWorkspaceRequestDto {
	urlName: string
	title: string
	/** @nullable */
	icon?: string | null
	assigneeStatusEditable?: boolean
	chatNotification?: boolean
	mailNotification?: boolean
	pikudId: number
	/** List of managers' emails for the workspace request */
	managers: string[]
}